refactor : move rating fetch method QUZ-113
