+++ matching-service/matching-domain/src/main/kotlin/com/grepp/quizy/matching/domain/game/GameFetcher.kt
@@ -5,4 +5,6 @@ import com.grepp.quizy.matching.domain.user.UserId
 
 interface GameFetcher {
     fun requestGameRoomId(userIds: List<UserId>, subject: InterestCategory): GameRoomId
+
+    fun requestGameRating(userId: UserId): GameRating
 }
